backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
crewai[google-genai]>=0.1.0
firecrawl-py>=0.0.1
pyyaml>=6.0
requests>=2.31.0
